Adding `strip` to the avatar on relative files
In testing I found that the avatar resolution when referring to assets was failing out before it hit the img tag because the `relative_url` could not resolve `<whatever>/<lots of space>\n<more space>\n`. Adding a `| strip` to the resolution fixes it.  The other cases get the same behavior later.