refactor(contact): Adds conditional for unset contact options in _config.yml.
For GitHub, GitLab, and Twitter, they will be displayed, regardless of a username being set or not. Updates the conditional for capture/assign of `url` so that it is `nil` if no username was provided. Then when the check for `url` is done for building the link, it will be skipped if no username is configured.