refactor(sidebar contact): Adds a conditional to check if a GitHub or Twitter username has been configured before displaying the icon in as a contact.
Ensures a GitHub or Twitter username is configured in `_config.yml` before displaying them in the sidebar. Adds a conditional to capture/assign `url` so that it is `nil` if no username was provided. That way when the check for `url` is done for building the link, it will be skipped if no username is configured.